chore(replay): extract BlockProxy and RequestInterceptor from PlayerController
Split the monolithic PlayerController into focused modules:
- BlockProxy: owns block list fetching and recording-api proxying
- RequestInterceptor: centralizes all Puppeteer request interception
  (player HTML, blocks, sub-frame stylesheets, media abort)
- PlayerController: now only handles player communication and state

PlayerController owns the RequestInterceptor and calls install() inside
load(), making the ordering constraint (interception before navigation)
invisible to callers. capturePlayback takes just the player and uses
player.waitForSettled() for the CDP beginFrame gate.
