+++ nodejs/src/session-replay/recording-rasterizer/capture/recorder.ts
@@ -1,99 +1,19 @@
-import * as fs from 'fs/promises'
-import { Page } from 'puppeteer'
-
 import { config as defaultConfig } from '../config'
-import { RasterizationError } from '../errors'
 import { type Logger, createLogger } from '../logger'
 import { RasterizeRecordingInput, RecordingResult } from '../types'
 import { elapsed } from '../utils'
-import { AssetProxy } from './asset-proxy'
+import { BlockProxy } from './block-proxy'
 import { BrowserPool } from './browser-pool'
-import { buildCaptureConfig, capturePlayback } from './capture'
-import { PlayerController, buildPlayerConfig, fetchBlockList } from './player'
-
-export const playerHtmlCache = {
-    _html: null as string | null,
-
-    async load(path?: string): Promise<string> {
-        const htmlPath = path || defaultConfig.playerHtmlPath
-        this._html = await fs.readFile(htmlPath, 'utf-8')
-        return this._html
-    },
-
-    get(): string {
-        if (!this._html) {
-            throw new Error('Player HTML not loaded — call playerHtmlCache.load() before recording')
-        }
-        return this._html
-    },
-
-    reset(): void {
-        this._html = null
-    },
-}
-
-export function validateInput(input: RasterizeRecordingInput): void {
-    if (!input.session_id) {
-        throw new RasterizationError('session_id is required', false, 'INVALID_INPUT')
-    }
-    if (!input.team_id || input.team_id <= 0) {
-        throw new RasterizationError('team_id must be a positive integer', false, 'INVALID_INPUT')
-    }
-    if (input.playback_speed !== undefined && input.playback_speed <= 0) {
-        throw new RasterizationError(
-            `playback_speed must be positive, got: ${input.playback_speed}`,
-            false,
-            'INVALID_INPUT'
-        )
-    }
-    if (input.capture_timeout != null && input.capture_timeout <= 0) {
-        throw new RasterizationError(
-            `capture_timeout must be positive, got: ${input.capture_timeout}`,
-            false,
-            'INVALID_INPUT'
-        )
-    }
-    if (input.recording_fps !== undefined && input.recording_fps <= 0) {
-        throw new RasterizationError(
-            `recording_fps must be positive, got: ${input.recording_fps}`,
-            false,
-            'INVALID_INPUT'
-        )
-    }
-    if (input.trim != null && input.trim <= 0) {
-        throw new RasterizationError(`trim must be positive, got: ${input.trim}`, false, 'INVALID_INPUT')
-    }
-}
-
-/**
- * Prepare a browser page for capture: set the viewport and optionally
- * wire up browser log forwarding.
- */
-async function preparePage(
-    page: Page,
-    viewport: { width: number; height: number },
-    captureLogs: boolean,
-    log: Logger
-): Promise<void> {
-    if (captureLogs) {
-        const browserLog = log.child({ source: 'browser' })
-        page.on('console', (msg) => {
-            const level = msg.type() === 'error' ? 'error' : msg.type() === 'warn' ? 'warn' : 'info'
-            browserLog[level](msg.text())
-        })
-        page.on('pageerror', (err) => browserLog.error({ type: 'pageerror' }, (err as Error).message))
-        page.on('requestfailed', (req) =>
-            browserLog.error({ type: 'requestfailed', url: req.url() }, req.failure()?.errorText || 'unknown')
-        )
-    }
-
-    await page.setViewport({ width: viewport.width, height: viewport.height, deviceScaleFactor: 1 })
-}
+import { capturePlayback } from './capture'
+import { CapturePage } from './capture-page'
+import { buildCaptureConfig, buildPlayerConfig, validateInput } from './config'
+import { PlayerController } from './player'
 
 export async function rasterizeRecording(
     pool: BrowserPool,
     input: RasterizeRecordingInput,
     outputPath: string,
+    playerHtml: string,
     cfg: typeof defaultConfig = defaultConfig,
     log: Logger = createLogger({ session_id: input.session_id, team_id: input.team_id }),
     onProgress?: () => void
@@ -101,26 +21,34 @@ export async function rasterizeRecording(
     validateInput(input)
 
     const setupStart = process.hrtime()
-    const baseHtml = playerHtmlCache.get()
     const captureConfig = buildCaptureConfig(input)
 
-    const page = await pool.getPage()
+    const rawPage = await pool.getPage()
     let player: PlayerController | null = null
     try {
         const viewport = {
             width: input.viewport_width || 1280,
             height: input.viewport_height || 720,
         }
-        await preparePage(page, viewport, cfg.captureBrowserLogs, log)
+        const playerUrl = `${cfg.siteUrl}/player`
+        const capturePage = await CapturePage.prepare(
+            rawPage,
+            viewport,
+            playerUrl,
+            playerHtml,
+            cfg.captureBrowserLogs,
+            log
+        )
 
-        const assetProxy = new AssetProxy(page, log)
-        player = new PlayerController(page, baseHtml, cfg, assetProxy, log)
-        const blocks = await fetchBlockList(input, cfg)
-        log.info({ blockCount: blocks.length }, 'block listing fetched')
-        const playerConfig = buildPlayerConfig(input, captureConfig.playbackSpeed, blocks.length)
+        const blockProxy = new BlockProxy(cfg, log)
+        const blockCount = await blockProxy.fetchBlocks(input)
+        log.info({ blockCount }, 'block listing fetched')
+
+        const playerConfig = buildPlayerConfig(input, captureConfig.playbackSpeed, blockCount)
+        player = new PlayerController(capturePage, blockProxy, log)
 
         log.info('loading player')
-        await player.load(playerConfig, blocks)
+        await player.load(playerConfig)
         log.info('player loaded, waiting for recording data')
 
         await player.waitForStart(playerConfig)
@@ -129,15 +57,7 @@ export async function rasterizeRecording(
 
         const setupS = elapsed(setupStart)
 
-        const captureResult = await capturePlayback(
-            page,
-            player,
-            assetProxy,
-            captureConfig,
-            outputPath,
-            log,
-            onProgress
-        )
+        const captureResult = await capturePlayback(player, captureConfig, outputPath, log, onProgress)
 
         return {
             video_path: outputPath,
@@ -151,6 +71,6 @@ export async function rasterizeRecording(
         }
     } finally {
         player?.dispose()
-        await pool.releasePage(page)
+        await pool.releasePage(rawPage)
     }
 }
