+++ nodejs/src/session-replay/recording-rasterizer/temporal/worker.ts
@@ -6,7 +6,7 @@ import * as prometheus from 'prom-client'
 import express from 'ultimate-express'
 
 import { BrowserPool } from '../capture/browser-pool'
-import { playerHtmlCache } from '../capture/recorder'
+import { playerHtmlCache } from '../capture/capture-page'
 import { config } from '../config'
 import { createLogger } from '../logger'
 import { createActivities } from './activities'
@@ -116,14 +116,14 @@ async function main(): Promise<void> {
 
     log.info('browser pool launched')
 
-    await playerHtmlCache.load()
+    const playerHtml = await playerHtmlCache.load()
     log.info({ path: config.playerHtmlPath }, 'player html loaded')
 
     const worker = await Worker.create({
         connection,
         namespace: config.temporalNamespace,
         taskQueue: config.taskQueue,
-        activities: createActivities(pool),
+        activities: createActivities(pool, playerHtml),
         maxConcurrentActivityTaskExecutions: config.maxConcurrentActivities,
         dataConverter: config.secretKey ? { payloadCodecs: [new EncryptionCodec(config.secretKey)] } : undefined,
     })
